// src/services/api/auth/register/register-payload.ts
export type RegisterPayload = {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  gender: string;
  dateOfBirth: string;
  address: string;
  phoneNumber: string;
  isTeacherRegistration: boolean;
  bio: string;
  department: string;
  parentPhoneNumber: string;
  grade: string;
  school: string;
};